Bring the struct Event out of C3DFileAdapter class.
